backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
Pillow>=9.5.0
django-cors-headers>=4.0.0
django-admin-sortable2>=2.1.8
djangorestframework-simplejwt>=5.2.2
python-social-auth>=0.3.6
django-filter>=23.2
